+++ tap_iterable/iterable.py
@@ -10,7 +10,7 @@ import backoff
 import requests
 import logging
 import tap_iterable.helper as helper
-from tap_iterable.exceptions import IterableRateLimitError, IterableNotAvailableError, \
+from tap_iterable.exceptions import IterableRateLimitError, IterableNotAvailableError, IterableServer5xxError, \
   raise_for_error
 
 LOGGER = logging.getLogger()
@@ -43,11 +43,12 @@ class Iterable(object):
     else:
       yield strptime_with_tz(start_date).strftime("%Y-%m-%d %H:%M:%S")
 
-  @backoff.on_exception(backoff.constant,
-                        (IterableRateLimitError, IterableNotAvailableError),
+  @backoff.on_exception(backoff.expo,
+                        (IterableRateLimitError, IterableNotAvailableError, IterableServer5xxError),
+                        max_tries=7,
                         jitter=None,
-                        interval=30,
-                        max_tries=5)
+                        base=2,
+                        factor=2)
   def _get(self, path, stream=True, **kwargs):
     """ The actual `get` request.  """
     uri = "{uri}{path}".format(uri=self.uri, path=path)
@@ -77,6 +78,13 @@ class Iterable(object):
   def get_user_fields(self):
     """Get custom user fields, used for generating `users` schema in `discover`."""
     return self.get("users/getFields")
+
+
+  def check_api_credentials(self):
+    """Checks if the provided API credentials are valid by making a test request."""
+    LOGGER.info("Checking API credentials")
+    self.get("channels", limit=1)
+    LOGGER.info("API credentials are valid")
 
 
   #
@@ -141,17 +149,29 @@ class Iterable(object):
     ]
     # `templates` API bug where it doesn't extract the records 
     #  where `startDateTime`= 2023-03-01+07%3A31%3A15 though record exists
-    #  hence, substracting one second so that we could extract atleast one record 
-    bookmark_val = strptime_with_tz(bookmark) - timedelta(seconds=1)
+    #  hence, substracting one second so that we could extract atleast one record
+
+    # ################################ Template API Issue WORKAROUND #################################
+    # We are moving with pseudo-incremental logic here to avoid missing out records due to API limitations.
+
+    # Reason for the workaround:
+    # -------------------------
+    # Iterable’s Templates API is not designed to support update-based incremental syncs.
+    # While it does have an updatedAt field, the API does not allow filtering based on it.
+    # Using updatedAt as a bookmark key will have a possibility of missing out records as API will always filter on createdAt.
+    # Even very old records if updated recently will be missed out as createdAt will always be old and we are never passing that created on to API.
+
+    ########################################################################################
+
+    bookmark_val = strptime_with_tz(bookmark)
     bookmark = strftime(bookmark_val)
     for template_type in template_types:
       for medium in message_mediums:
-        for kwargs in self.get_start_end_date(bookmark):
-          res = self.get("templates", templateTypes=template_type, messageMedium=medium, **kwargs)
-          for t in res["templates"]:
-            rec_date_time = strptime_with_tz(helper.epoch_to_datetime_string(t[column_name]))
-            if rec_date_time >= bookmark_val:
-              yield t
+        res = self.get("templates", templateType=template_type, messageMedium=medium)
+        for t in res["templates"]:
+          rec_date_time = strptime_with_tz(helper.epoch_to_datetime_string(t[column_name]))
+          if rec_date_time >= bookmark_val:
+            yield t
 
 
   def metadata(self, column_name=None, bookmark=None):
@@ -182,4 +202,3 @@ class Iterable(object):
       def get_data():
         return self._get("export/data.json", dataTypeName=data_type_name, **kwargs), kwargs['endDateTime']
       yield get_data
-